feat(): substrate types organisation
