+++ packages/beacon-blockchain-substrate/src/index.ts
@@ -1,1 +1,9 @@
 export { SubstrateBlockchain } from './blockchain'
+export * from './types/message-type'
+export * from './types/permission-scope'
+export * from './types/messages/permission-request'
+export * from './types/messages/permission-response'
+export * from './types/messages/sign-payload-request'
+export * from './types/messages/sign-payload-response'
+export * from './types/messages/transfer-request'
+export * from './types/messages/transfer-response'
