--- packages/beacon-blockchain-substrate/src/messages/permission-request.ts
@@ -1,1 +1,0 @@
-// This should be the same for all protocols
