+++ packages/beacon-blockchain-substrate/src/types/message-type.ts
@@ -1,0 +1,4 @@
+export enum SubstrateMessageType {
+  'transfer_request' = 'transfer_request',
+  'sign_payload_request' = 'sign_payload_request'
+}
