+++ packages/beacon-blockchain-substrate/src/types/messages/sign-payload-response.ts
@@ -1,0 +1,13 @@
+export type SubstrateSignPayloadResponse =
+  | {
+      transactionHash: string
+    }
+  | {
+      transactionHash: string
+      signature: string
+      payload?: string
+    }
+  | {
+      signature: string
+      payload?: string
+    }
