+++ packages/beacon-blockchain-substrate/src/types/messages/transfer-response.ts
@@ -1,0 +1,13 @@
+export type SubstrateTransferResponse =
+  | {
+      transactionHash: string
+    }
+  | {
+      transactionHash: string
+      signature: string
+      payload?: string
+    }
+  | {
+      signature: string
+      payload?: string
+    }
