+++ packages/beacon-blockchain-substrate/src/types/permission-scope.ts
@@ -1,0 +1,5 @@
+export enum SubstratePermissionScope {
+  'transfer' = 'transfer',
+  'sign_payload_json' = 'sign_payload_json',
+  'sign_payload_raw' = 'sign_payload_raw'
+}
